+++ packages/components/molecules/f-card-with-content/test/accessibility/axe-accessibility.spec.js
@@ -1,6 +1,6 @@
 import { getAccessibilityTestResults } from '../../../../../../test/utils/axe-helper';
 
-const CardWithContent = require('../../test-utils/component-objects/f-cardWithContent.component');
+const CardWithContent = require('../../test-utils/component-objects/f-card-with-content.component');
 
 const cardWithContent = new CardWithContent();
 
@@ -9,9 +9,9 @@ describe('Accessibility tests', () => {
         cardWithContent.load();
     });
 
-    it('a11y - should test f-cardWithContent component WCAG compliance', () => {
+    it('a11y - should test f-card-with-content component WCAG compliance', () => {
         // Act
-        const axeResults = getAccessibilityTestResults('f-cardWithContent');
+        const axeResults = getAccessibilityTestResults('f-card-with-content');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);