+++ packages/components/organisms/f-status-banner/test/accessibility/axe-accessibility.spec.js
@@ -1,6 +1,6 @@
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 
-const StatusBanner = require('../../test-utils/component-objects/f-statusBanner.component');
+const StatusBanner = require('../../test-utils/component-objects/f-status-banner.component');
 
 let statusBanner;
 
@@ -9,9 +9,9 @@ describe('Accessibility tests', () => {
         statusBanner = new StatusBanner();
         statusBanner.load();
     });
-    it('a11y - should test f-statusBanner component WCAG compliance', () => {
+    it('a11y - should test f-status-banner component WCAG compliance', () => {
         // Act
-        const axeResults = getAccessibilityTestResults('f-statusBanner');
+        const axeResults = getAccessibilityTestResults('f-status-banner');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);